trying to sell drums

+++ week_07/day_5/week_7_music_shop_HW/src/main/java/Interface/ISell.java
@@ -2,6 +2,6 @@ package Interface;
 
 public interface ISell {
 
-    public int sell(int selling);
+    public double sell(double selling);
 
 }
